Use mkhelp for Makefile help shortcut
Update the docstring syntax to what is expected by mkhelp. This paves
the way for using mkhelp to parse docstrings for use by the makefile
provider that will be added.

Also move workflows defined as aliases and invoked from make to avoid
circular dependencies when the makefile provider is added to kindly.