Remove greetings provider
This was good for testing before I had any useful providers but now it
only gets in the way of using the program since it provides one extra,
useless command.

Approximate the greeting provider with an alias because this command
was more fun than the previous greeting alias and to demonstrate that
more advances scripts can be aliased (it is clearly messy though and
should probably not be done).